chore: temporary disabled starter screen (#3663)
* chore: temporary disabeld starter screen

* chore: create useStarter screen

* chore: fix typo

* chore: update config starter screen

* chore: update comment code

* chore: renmove some commented code
